Begin transition to a more modularized language.
Moved the flowchart model/ast to a universal OWL file. The idea is that
instead of having every new parser redefine the model in its
implementing language, every parser uses the OWL model and parses to
OWL. Since all flowcharts will be represented in OWL, there can be a
single program (or a separate class of programs) that parses a
flowchart encoded in OWL to a visual representation.

The Java parser has been rewritten to produce flowcharts in OWL (using
Jena); the ‘universal’ flowchart creator still needs to be
written—probably with Graphviz still, maybe with a different library.